Split FetchCtrl into many other controllers; other big changes.
There's also a 'states' service to manage state between the controllers.

And a 'conversion' one to handle LTc exchange rates.

The server now submits news only on fetch or when a user joins.
There is no longer a way to coax the server for an update or to fetch.

Now I should be able to change the layout significantly!